Langchain/requirements.txt
python-dotenv
langchain
langchain-core
langchain-community
langchain-groq
langchain-huggingface
langchain-text-splitters
groq
fastembed
faiss-cpu
selenium
webdriver-manager
huggingface-hub
beautifulsoup4
requests
urllib3